(refactor) Rename files to match naming conventions
