+++ packages/esm-admin-openconceptlab-app/src/import/import.component.tsx
@@ -1,4 +1,4 @@
-import { showNotification } from '@openmrs/esm-framework';
+import React, { useCallback, useEffect, useState } from 'react';
 import {
   Button,
   ButtonSkeleton,
@@ -12,10 +12,10 @@ import {
   Stack,
   SkeletonText,
 } from '@carbon/react';
-import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import styles from './import.component.scss';
+import { showNotification } from '@openmrs/esm-framework';
 import { startImportWithFile, startImportWithSubscription, useSubscription } from './import.resource';
+import styles from './import.scss';
 
 const Import: React.FC = () => {
   const { t } = useTranslation();
