+++ packages/esm-admin-openconceptlab-app/src/previous-imports/import-overview/import-overview.test.tsx
@@ -3,7 +3,6 @@ import { screen, waitFor } from '@testing-library/react';
 import { formatDatetime } from '@openmrs/esm-framework';
 import { renderWithSwr } from '../../../../../tools/test-helpers';
 import { mockPreviousImports } from '../../../../../__mocks__/openconceptlab.mock';
-import { Import } from '../../types';
 import ImportOverview from './import-overview.component';
 
 describe(`Import Overview component`, () => {
@@ -72,7 +71,7 @@ describe(`Import Overview component`, () => {
   });
 });
 
-function renderImportOverviewComponent(importObject: Import) {
+function renderImportOverviewComponent(importObject) {
   renderWithSwr(<ImportOverview selectedImportObject={importObject} />);
 }
