+++ packages/esm-admin-openconceptlab-app/src/root.scss
