+++ packages/esm-admin-openconceptlab-app/src/subscription/subscription.component.tsx
@@ -1,4 +1,5 @@
-import { showNotification } from '@openmrs/esm-framework';
+import React, { useCallback, useEffect, useState } from 'react';
+import { useSWRConfig } from 'swr';
 import {
   Button,
   ButtonSkeleton,
@@ -13,12 +14,11 @@ import {
   TextInput,
   TextInputSkeleton,
 } from '@carbon/react';
-import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { showNotification } from '@openmrs/esm-framework';
 import { deleteSubscription, updateSubscription, useSubscription } from './subscription.resource';
-import styles from './subscription.component.scss';
-import { useSWRConfig } from 'swr';
 import { isVersionDefinedInUrl } from '../utils';
+import styles from './subscription.scss';
 
 const Subscription: React.FC = () => {
   const { t } = useTranslation();
