+++ packages/esm-admin-openconceptlab-app/src/subscription/subscription.resource.ts
@@ -1,7 +1,7 @@
 import { openmrsFetch } from '@openmrs/esm-framework';
 import useSWR from 'swr';
-import { Subscription } from '../types';
 import isNil from 'lodash-es/isNil';
+import type { Subscription } from '../types';
 
 export function useSubscription() {
   const { data, error, isValidating } = useSWR<{ data: { results: Subscription[] } }, Error>(
