--- packages/esm-system-admin-app/src/root.component.scss
@@ -1,23 +1,0 @@
-@use "@carbon/styles/scss/spacing";
-@use "@carbon/styles/scss/type";
-@import "~@openmrs/esm-styleguide/src/vars";
-
-.main {
-  background-color: $ui-02;
-}
-
-.moduleHeader {
-  @include type.type-style('productive-heading-03');
-  padding-left: spacing.$spacing-05;
-  padding-top: spacing.$spacing-05;
-  padding-bottom: spacing.$spacing-05;
-}
-
-.tabList {
-  padding: 0 spacing.$spacing-05;
-}
-
-.tabPanel {
-  height: 100%;
-  background-color: $ui-01;
-}
